Clean up component hierarchy
