--- front/app/components/admin/PostManager/components/LazyPostPreview/Idea/LazyAdminIdeaEdit.tsx
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export default lazy(() => import('./AdminIdeaEdit'));
