--- front/app/components/admin/PostManager/components/LazyPostPreview/Idea/LazyIdeaContent.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export default lazy(() => import('./AdminIdeaContent'));
