--- front/app/components/admin/PostManager/components/LazyPostPreview/index.tsx
@@ -1,8 +1,0 @@
-import { lazy } from 'react';
-
-export default lazy(
-  () =>
-    import(
-      'components/admin/PostManager/components/LazyPostPreview/PostPreview'
-    )
-);
