+++ front/app/components/admin/PostManager/components/PostPreview/Idea/FeedbackSettings.tsx
