+++ front/app/components/admin/PostManager/components/PostPreview/Idea/ReactionPreview.tsx
