+++ front/app/components/admin/PostManager/components/PostPreview/Initiative/FeedbackSettings.tsx
